Intials view in customer table view cell
