--- ReSwiftTest/LogoPlaceholderView.xib
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="11762" systemVersion="16C67" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="11757"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB" customClass="LogoPlaceholderView" customModule="ReSwiftTest" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="50" height="50"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="g6f-0w-v7K">
-                    <rect key="frame" x="0.0" y="0.0" width="50" height="50"/>
-                    <color key="backgroundColor" red="0.41960784313725491" green="0.17254901960784313" blue="0.95294117647058818" alpha="1" colorSpace="calibratedRGB"/>
-                </view>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="AB" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="lVg-mj-tc1">
-                    <rect key="frame" x="10" y="12" width="30.5" height="26.5"/>
-                    <fontDescription key="fontDescription" type="boldSystem" pointSize="22"/>
-                    <color key="textColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                    <nil key="highlightedColor"/>
-                </label>
-            </subviews>
-            <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="calibratedWhite"/>
-            <constraints>
-                <constraint firstAttribute="bottom" secondItem="g6f-0w-v7K" secondAttribute="bottom" id="8lF-hI-tZk"/>
-                <constraint firstItem="g6f-0w-v7K" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" id="Ewu-FH-QoO"/>
-                <constraint firstItem="g6f-0w-v7K" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" id="Mtu-Nq-tu0"/>
-                <constraint firstAttribute="trailing" secondItem="g6f-0w-v7K" secondAttribute="trailing" id="On0-nO-8Ye"/>
-                <constraint firstItem="lVg-mj-tc1" firstAttribute="centerX" secondItem="iN0-l3-epB" secondAttribute="centerX" id="PFY-B9-kjY"/>
-                <constraint firstItem="lVg-mj-tc1" firstAttribute="centerY" secondItem="iN0-l3-epB" secondAttribute="centerY" id="UeO-Pu-gnY"/>
-            </constraints>
-            <nil key="simulatedStatusBarMetrics"/>
-            <nil key="simulatedTopBarMetrics"/>
-            <nil key="simulatedBottomBarMetrics"/>
-            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
-            <connections>
-                <outlet property="circleView" destination="g6f-0w-v7K" id="0kM-WE-ixz"/>
-                <outlet property="initialsLabel" destination="lVg-mj-tc1" id="RVP-rh-ryk"/>
-            </connections>
-            <point key="canvasLocation" x="-38" y="-128"/>
-        </view>
-    </objects>
-</document>
